Refactor plugins dependencies
