+++ src/common.ts
@@ -1,0 +1,7 @@
+/**
+ * Modules which plugins must dependents
+ * Separating each dependency structure to let the browser reduce dependencies and memory capacity
+ */
+export { Request } from './engine/request';
+export { Context } from './engine/context';
+export { Struct } from './engine/struct';
