+++ src/plugins/index.ts
@@ -1,4 +1,8 @@
+/**
+ * Export plugins for Node/Browser.
+ */
 export * from './addTimeNow';
 export * from './noReplyHandle';
 export * from './normalize';
 export * from './nlu';
+export * from '../version';
